LIN bus driver & UART set changed
Working on it. Finish with the commands:
- SSPDLIN
- GSPDLIN

To set up the baud rate it was defined with short int but I needed a long int, so I changed this definition.

+++ UART.h
@@ -32,7 +32,7 @@
 
 
 extern void uart_init(unsigned char uartx);
-extern void uart_set(unsigned char uartx, unsigned char parity, unsigned char stopBit, unsigned char polarity, unsigned char autobaud, unsigned short int baudrate);
+extern void uart_set(unsigned char uartx, unsigned char parity, unsigned char stopBit, unsigned char polarity, unsigned char autobaud, unsigned long int baudrate);
 extern void uart_InitInterrupt(unsigned char uartx, unsigned char action);
 extern void uart_enable(unsigned char uartx);
 extern void uart_disable(unsigned char uartx);
